rename test folders
